+++ src/front/src/components/screens/ScreenSpawnPortal.tsx
@@ -93,7 +93,7 @@ export const ScreenSpawnPortal = () => {
                                 value={amount}
                                 onChange={(e) =>
                                     setAmount(
-                                        parseInt(
+                                        parseFloat(
                                             e.target.value
                                                 ? e.target.value
                                                 : "0",
