+++ src/front/src/components/screens/ScreenWithdraw.tsx
@@ -101,7 +101,7 @@ export const ScreenWithdraw = () => {
                                 value={amount}
                                 onChange={(e) =>
                                     setAmount(
-                                        parseInt(
+                                        parseFloat(
                                             e.target.value
                                                 ? e.target.value
                                                 : "0",
